Make architeckture filter module
