+++ app/src/main/java/ru/practicum/android/diploma/root/App.kt
@@ -7,7 +7,9 @@ import ru.practicum.android.diploma.data.db.di.dbDataModule
 import ru.practicum.android.diploma.data.networkclient.di.networkDataModule
 import ru.practicum.android.diploma.data.sp.di.spDataModule
 import ru.practicum.android.diploma.favorites.di.favoritesModule
-import ru.practicum.android.diploma.filter.di.filterModule
+import ru.practicum.android.diploma.filter.filter.di.filterModule
+import ru.practicum.android.diploma.filter.place.di.placeModule
+import ru.practicum.android.diploma.filter.profession.di.professionModule
 import ru.practicum.android.diploma.search.di.searchModule
 import ru.practicum.android.diploma.team.di.teamModule
 import ru.practicum.android.diploma.vacancy.di.vacancyDetailModule
@@ -25,6 +27,8 @@ class App : Application() {
                 dbDataModule,
                 searchModule,
                 filterModule,
+                professionModule,
+                placeModule,
                 favoritesModule,
                 vacancyDetailModule,
                 teamModule,
