--- feature-filter/src/main/java/ru/practicum/android/diploma/filter/domain/model/AreaInReference.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.filter.domain.model
-
-data class AreaInReference(
-    val areas: List<AreaInReference>,
-    val id: String,
-    val name: String,
-    val parentId: String?,
-)
