--- feature-filter/src/main/java/ru/practicum/android/diploma/filter/domain/model/Industry.kt
@@ -1,6 +1,0 @@
-package ru.practicum.android.diploma.filter.domain.model
-
-data class Industry(
-    val id: String,
-    val name: String,
-)
