+++ feature-filter/src/main/java/ru/practicum/android/diploma/filter/filter/di/professionModule.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.filter.filter.di
+
+import org.koin.dsl.module
+
+val filterModule = module {
+
+}
