+++ feature-filter/src/main/java/ru/practicum/android/diploma/filter/place/domain/model/AreaInReference.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.filter.place.domain.model
+
+data class AreaInReference(
+    val areas: List<AreaInReference>,
+    val id: String,
+    val name: String,
+    val parentId: String?,
+)
