+++ feature-filter/src/main/java/ru/practicum/android/diploma/filter/place/domain/model/Country.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.filter.place.domain.model
+
+data class Country(val id: String, val name: String)
