+++ feature-filter/src/main/java/ru/practicum/android/diploma/filter/place/domain/repository/RegionRepository.kt
@@ -1,0 +1,9 @@
+package ru.practicum.android.diploma.filter.place.domain.repository
+
+import kotlinx.coroutines.flow.Flow
+import ru.practicum.android.diploma.commonutils.Resource
+import ru.practicum.android.diploma.filter.place.domain.model.AreaInReference
+
+internal interface RegionRepository {
+    fun listAreas(): Flow<Resource<List<AreaInReference>>>
+}
