+++ feature-filter/src/main/java/ru/practicum/android/diploma/filter/place/presentation/viewmodel/state/RegionState.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.filter.place.presentation.viewmodel.state
+
+sealed interface RegionState {
+    data class Content(val regions: List<Map<String, String>>) : RegionState
+    object Empty : RegionState
+    object Error : RegionState
+}
