--- feature-filter/src/main/java/ru/practicum/android/diploma/filter/presentation/ui/PlaceFragment.kt
@@ -1,42 +1,0 @@
-package ru.practicum.android.diploma.filter.presentation.ui
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-import ru.practicum.android.diploma.filter.R
-import ru.practicum.android.diploma.filter.databinding.FragmentPlaceBinding
-
-internal class PlaceFragment : Fragment() {
-    private var _binding: FragmentPlaceBinding? = null
-    private val binding get() = _binding!!
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        _binding = FragmentPlaceBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.country.setOnClickListener {
-            findNavController().navigate(R.id.action_placeFragment_to_countryFragment)
-        }
-        binding.region.setOnClickListener {
-            findNavController().navigate(R.id.action_placeFragment_to_regionFragment)
-        }
-        binding.buttonLeftPlace.setOnClickListener {
-            findNavController().navigateUp()
-        }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
-}
