--- feature-filter/src/main/java/ru/practicum/android/diploma/filter/presentation/ui/RegionFragment.kt
@@ -1,35 +1,0 @@
-package ru.practicum.android.diploma.filter.presentation.ui
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-import ru.practicum.android.diploma.filter.databinding.FragmentRegionBinding
-
-internal class RegionFragment : Fragment() {
-    private var _binding: FragmentRegionBinding? = null
-    private val binding get() = _binding!!
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        _binding = FragmentRegionBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.buttonLeftRegion.setOnClickListener {
-            findNavController().navigateUp()
-        }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
-}
