--- feature-filter/src/main/java/ru/practicum/android/diploma/filter/presentation/viewmodel/ProfessionViewModel.kt
@@ -1,14 +1,0 @@
-package ru.practicum.android.diploma.filter.presentation.viewmodel
-
-import android.app.Application
-import androidx.lifecycle.AndroidViewModel
-
-class ProfessionViewModel(
-    private val application: Application,
-) : AndroidViewModel(application) {
-
-    companion object {
-        private const val TAG: String = "ProfessionViewModel"
-    }
-
-}
