--- feature-filter/src/main/java/ru/practicum/android/diploma/filter/presentation/viewmodel/state/Country.kt
@@ -1,3 +1,0 @@
-package ru.practicum.android.diploma.filter.presentation.viewmodel.state
-
-data class Country(val id: String, val name: String)
