--- feature-filter/src/main/java/ru/practicum/android/diploma/filter/presentation/viewmodel/state/CountryState.kt
@@ -1,7 +1,0 @@
-package ru.practicum.android.diploma.filter.presentation.viewmodel.state
-
-sealed interface CountryState {
-    data class Content(val coutries: List<Country>) : CountryState
-    object Empty : CountryState
-    object Error : CountryState
-}
