--- feature-filter/src/main/java/ru/practicum/android/diploma/filter/presentation/viewmodel/state/RegionState.kt
@@ -1,7 +1,0 @@
-package ru.practicum.android.diploma.filter.presentation.viewmodel.state
-
-sealed interface RegionState {
-    data class Content(val regions: List<Map<String, String>>) : RegionState
-    object Empty : RegionState
-    object Error : RegionState
-}
