+++ feature-filter/src/main/java/ru/practicum/android/diploma/filter/profession/domain/model/Industry.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.filter.profession.domain.model
+
+data class Industry(
+    val id: String,
+    val name: String,
+)
