+++ feature-filter/src/main/java/ru/practicum/android/diploma/filter/profession/presentation/viewmodel/ProfessionViewModel.kt
@@ -1,0 +1,14 @@
+package ru.practicum.android.diploma.filter.profession.presentation.viewmodel
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+
+class ProfessionViewModel(
+    private val application: Application,
+) : AndroidViewModel(application) {
+
+    companion object {
+        private const val TAG: String = "ProfessionViewModel"
+    }
+
+}
