+++ feature-filter/src/main/res/layout/fragment_country.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".presentation.ui.CountryFragment">
+    tools:context=".place.presentation.ui.CountryFragment">
 
     <ImageView
         android:id="@+id/buttonLeftCountry"
