+++ feature-filter/src/main/res/layout/fragment_filter.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".presentation.ui.FilterFragment">
+    tools:context=".filter.presentation.ui.FilterFragment">
 
     <ImageView
         android:id="@+id/buttonLeftFilter"
