+++ feature-filter/src/main/res/layout/fragment_place.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".presentation.ui.PlaceFragment">
+    tools:context=".place.presentation.ui.PlaceFragment">
     <ImageView
         android:id="@+id/buttonLeftPlace"
         style="@style/ButtonLeft"
