+++ feature-filter/src/main/res/layout/fragment_profession.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".presentation.ui.SearchFragment">
+    tools:context=".profession.presentation.ui.SearchFragment">
 
     <!-- Guidelines -->
     <androidx.constraintlayout.widget.Guideline
