+++ feature-filter/src/main/res/navigation/filter_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/filterFragment"
-        android:name="ru.practicum.android.diploma.filter.presentation.ui.FilterFragment"
+        android:name="ru.practicum.android.diploma.filter.filter.presentation.ui.FilterFragment"
         android:label="FilterFragment">
         <action
             android:id="@+id/action_filterFragment_to_placeFragment"
@@ -18,7 +18,7 @@
     </fragment>
     <fragment
         android:id="@+id/placeFragment"
-        android:name="ru.practicum.android.diploma.filter.presentation.ui.PlaceFragment"
+        android:name="ru.practicum.android.diploma.filter.place.presentation.ui.PlaceFragment"
         tools:layout="@layout/fragment_place"
         android:label="fragment_place">
         <action
@@ -30,17 +30,17 @@
     </fragment>
     <fragment
         android:id="@+id/countryFragment"
-        android:name="ru.practicum.android.diploma.filter.presentation.ui.CountryFragment"
+        android:name="ru.practicum.android.diploma.filter.place.presentation.ui.CountryFragment"
         tools:layout="@layout/fragment_country"
         android:label="fragment_country" />
     <fragment
         android:id="@+id/regionFragment"
-        android:name="ru.practicum.android.diploma.filter.presentation.ui.RegionFragment"
+        android:name="ru.practicum.android.diploma.filter.place.presentation.ui.RegionFragment"
         tools:layout="@layout/fragment_region"
         android:label="fragment_region" />
     <fragment
         android:id="@+id/professionFragment"
-        android:name="ru.practicum.android.diploma.filter.presentation.ui.ProfessionFragment"
+        android:name="ru.practicum.android.diploma.filter.profession.presentation.ui.ProfessionFragment"
         tools:layout="@layout/fragment_profession"
         android:label="fragment_profession" />
 </navigation>
